\name{AIDSBlogs}
\alias{AIDSBlogs}
\docType{data}
\title{
The AIDSBlogs data set
}
\description{
The AIDS blog data set records the pattern of citation among 146 unique blogs related to AIDS patients and their support networks. The data were originally collected by Gopal (2007) <doi:10.1007/1-4020-5427-0_18> over a randomly selected three-day period in August 2005. The nodes in the network correspond to blogs and a directed edge from one blog to another indicates that the former had a link to the latter in their web page.
}
\usage{data("AIDSBlogs")}
\format{
A large network object, which can be managed with the network library, with 146 nodes.
}
\references{
Gopal, S., The evolving social geography of blogs, in Miller, H. J. (ed.),
Societies and Cities in the Age of Instant Access, The GeoJournal Library, vol. 88., pp. 275–293, 2007 <doi:10.1007/1-4020-5427-0_18>.
}
\examples{
data(AIDSBlogs)
}
\keyword{datasets}
