\name{amazonFineFoods}
\alias{amazonFineFoods}
\docType{data}
\title{
The Amazon Fine Foods data set
}
\description{
The Amazon Fine Foods data set has 1646 rows and 1735 columns, describing whether an user (row) has noted and reviewed a product (column) or not.
}
\usage{data("amazonFineFoods")}
\format{
  A data frame with binary values indicating whether an user (row) has noted and reviewed a product (column) or not.
}
\source{
https://snap.stanford.edu/data/web-FineFoods.html.
}
\examples{
data(amazonFineFoods)
}
\keyword{datasets}
