\name{Coworker}
\alias{Coworker}
\docType{data}
\title{
The Coworker data set from Lazega (2001)
}
\description{
Lazega (2001) <doi:10.2307/3556688> collected a network data set detailing interactions between a set of 71 lawyers in a corporate law firm in the USA. The data include measurements of the advice network, friendship network and co-worker network between the lawyers within the firm. Further covariates associated with each lawyer in the firm are also available including age, seniority, college education and office location.
}
\usage{data("Coworker")}
\format{
A large network object, which can be managed with the network library, with 71 nodes.
}
\references{
Lazega, E., The Collegial Phenomenon: The Social Mechanisms of Cooperation Among
Peers in a Corporate Law Partnership, Oxford University Press, 2001 <doi:10.2307/3556688>.
}
\examples{
data(Coworker)
}
\keyword{datasets}
