\name{credit}
\alias{credit}
\docType{data}
\title{
The Credit data set
}
\description{
The Credit data set has 66 rows and 11 columns, describing customers who took out loans from a credit company described with 11 categorical or ordinal variables.
}
\usage{data("credit")}
\format{
A data frame with 66 observations and 11 categorical or ordinal variables.
}
\source{
https://husson.github.io/data.html
}
\examples{
data(credit)
}
\keyword{datasets}
