\name{MBCbook-package}
\alias{MBCbook-package}
\alias{MBCbook}
\docType{package}
\title{
\packageTitle{MBCbook}
}
\description{
\packageDescription{MBCbook}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MBCbook}
\packageIndices{MBCbook}
}
\author{
\packageAuthor{MBCbook}

Maintainer: \packageMaintainer{MBCbook}
}
\references{
Charles Bouveyron and Gilles Celeux and T. Brendan Murphy and Adrian E. Raftery, Model-Based Clustering and Classification for Data Science: with Applications in R, Cambridge University Press, 2019.
}
\keyword{ package }