\name{puffin}
\alias{puffin}
\docType{data}
\title{
The puffin data set
}
\description{
The puffin data set contains 69 individuals (birds) described by 5 categorical variables, in addition to class labels.
}
\usage{data("puffin")}
\format{
A data frame with 69 observations and 6 variables.
  \describe{
    \item{\code{class}}{the class of the observations}
    \item{\code{gender}}{gender of the bird}
    \item{\code{eyebrow}}{gender of the bird}
    \item{\code{collar}}{gender of the bird}
    \item{\code{sub.caudal}}{gender of the bird}
    \item{\code{border}}{gender of the bird}
  }
}
\source{
The data were provided by Bretagnolle, V., Museum d'Histoire Naturelle, Paris.
}
\examples{
data(puffin)
}
\keyword{datasets}
