\name{UScongress}
\alias{UScongress}
\docType{data}
\title{
The US congress vote data set
}
\description{
The US congress vote data set contains the votes (yes, no, abstained or absent) of 434 members of the 98th US Congress on 16 different key issues. This data set involves three-level categorical data.
}
\usage{data("UScongress")}
\format{
  A data frame with 434 observations on 16 different key issues. The first variables indicates the political party of the congressmen.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Congressional+Voting+Records
}
\examples{
data(UScongress)
}
\keyword{datasets}
